Fix issue that incorrectly subscribed users in Campaign monitor
When null is passed to the core timestamp function, it returns the
current date. If the extension is in the default "timestamp" mode, it
will compare this timestamp to the last updated timestamp in CM. By
default Magento has a null value for the change status at attribute,
which means they are interpreted as todays date, instead of no date.
This change ensures we pass through an actual null time when no changed
status at date is available in Magento.